add config.go
