+++ cfg/config.yaml
@@ -1,0 +1,2 @@
+api:
+  gemini: ""
