--- gemini/cfg/config.yaml
@@ -1,2 +1,0 @@
-api:
-  gemini: "记得使用引号框起来"
