+++ go.mod
@@ -1,6 +1,11 @@
-module GinLearning
+module github.com/JoceHYJ/GinLearning
 
 go 1.20
+
+require (
+	github.com/gin-gonic/gin v1.9.1
+	gopkg.in/yaml.v2 v2.4.0
+)
 
 require (
 	github.com/bytedance/sonic v1.10.2 // indirect
@@ -8,7 +13,6 @@ require (
 	github.com/chenzhuoyu/iasm v0.9.1 // indirect
 	github.com/gabriel-vasile/mimetype v1.4.3 // indirect
 	github.com/gin-contrib/sse v0.1.0 // indirect
-	github.com/gin-gonic/gin v1.9.1 // indirect
 	github.com/go-playground/locales v0.14.1 // indirect
 	github.com/go-playground/universal-translator v0.18.1 // indirect
 	github.com/go-playground/validator/v10 v10.16.0 // indirect
